Fix mypy no-any-return errors in mission.py and permeability.py (#73)
* Fix mypy no-any-return errors in mission.py and permeability.py

Wrap sum(), min(), and max(min()) return values with float() to
satisfy mypy's no-any-return checks in functions declared to return
float.


* Fix mypy and ruff lint errors across swarm/ package

mypy fixes:
- Replace builtin `callable` with `typing.Callable` in quality.py
- Suppress import-untyped for requests in platforms.py
- Wrap response.json() with dict() for no-any-return in platforms.py
- Cast numpy bool/floating to Python bool/float in reflexivity.py
- Cast np.mean result to float in agents.py
- Add set[str] annotation for params_tested in agents.py
- Assign json.load to typed variable in workflow.py

ruff fixes (auto + manual):
- Sort import blocks (I001) across all 5 research files
- Remove unused imports (F401): json, datetime, unused re-exports
- Remove unused exception variable (F841)
- Remove extraneous f-string prefix (F541)
- Rename unused loop variables to _ prefix (B007)
- Add strict= to zip() call (B905)
- Replace dict comprehension with dict.fromkeys (C420)
- Rename ambiguous variable `l` to `lim` (E741)


---------

+++ src/env/mission.py
@@ -444,9 +444,9 @@ class MissionEconomy:
         metric = objective.target_metric
 
         if metric == "avg_p":
-            return sum(i.p for i in interactions) / len(interactions)
+            return float(sum(i.p for i in interactions)) / len(interactions)
         elif metric == "min_p":
-            return min(i.p for i in interactions)
+            return float(min(i.p for i in interactions))
         elif metric == "total_count":
             return float(len(interactions))
         elif metric == "acceptance_rate":
@@ -454,7 +454,7 @@ class MissionEconomy:
             return accepted / len(interactions)
         elif metric == "total_welfare":
             # Sum of expected surplus: p * s_plus - (1-p) * s_minus
-            return sum(i.p * 2.0 - (1 - i.p) * 1.0 for i in interactions)
+            return float(sum(i.p * 2.0 - (1 - i.p) * 1.0 for i in interactions))
         else:
             return 0.0
